+++ tests/functional/test_certification_task.py
@@ -18,21 +18,18 @@ import os
 import traceback
 import unittest
 
-import rally as rally_m
-from tests.check_samples import utils
+import rally_openstack
+from tests.functional import utils
 
 
 class TestPreCreatedTasks(unittest.TestCase):
-
-    def test_check_success(self):
-        rally = utils.Rally()
-        rally("deployment check")
 
     def test_task_samples_is_valid(self):
 
         rally = utils.Rally()
         full_path = os.path.join(
-            os.path.dirname(rally_m.__file__), os.pardir, "tasks", "openstack")
+            os.path.dirname(rally_openstack.__file__), os.pardir,
+            "tasks", "openstack")
         task_path = os.path.join(full_path, "task.yaml")
         args_path = os.path.join(full_path, "task_arguments.yaml")