+++ tests/unit/contexts/sahara/test_sahara_image.py
@@ -16,13 +16,13 @@ import mock
 
 from rally import exceptions
 from rally_openstack.contexts.sahara import sahara_image
+from tests.unit import fakes
 from tests.unit import test
 
 
 BASE_CTX = "rally.task.context"
 CTX = "rally_openstack.contexts.sahara.sahara_image"
 BASE_SCN = "rally.task.scenarios"
-SCN = "rally_openstack.scenarios"
 
 
 class SaharaImageTestCase(test.ScenarioTestCase):
@@ -43,7 +43,7 @@ class SaharaImageTestCase(test.ScenarioTestCase):
             for j in range(self.users_per_tenant):
                 self.users_key.append({"id": "%s_%s" % (str(i), str(j)),
                                        "tenant_id": str(i),
-                                       "credential": mock.MagicMock()})
+                                       "credential": fakes.FakeCredential()})
 
     @property
     def url_image_context(self):
@@ -60,7 +60,7 @@ class SaharaImageTestCase(test.ScenarioTestCase):
                     "username": "test_user"
                 }
             },
-            "admin": {"credential": mock.MagicMock()},
+            "admin": {"credential": fakes.FakeCredential()},
             "users": self.users_key,
             "tenants": self.tenants
         })
@@ -78,7 +78,7 @@ class SaharaImageTestCase(test.ScenarioTestCase):
                     "image_uuid": "some_id"
                 }
             },
-            "admin": {"credential": mock.MagicMock()},
+            "admin": {"credential": fakes.FakeCredential()},
             "users": self.users_key,
             "tenants": self.tenants,
         })
@@ -135,19 +135,17 @@ class SaharaImageTestCase(test.ScenarioTestCase):
             superclass=sahara_ctx.__class__,
             task_id=ctx["owner_id"])
 
-    @mock.patch("%s.glance.utils.GlanceScenario._create_image" % SCN,
-                return_value=mock.MagicMock(id=42))
     @mock.patch("%s.resource_manager.cleanup" % CTX)
     @mock.patch("%s.osclients.Clients" % CTX)
     def test_setup_and_cleanup_existing_image(
-            self, mock_clients, mock_cleanup,
-            mock_glance_scenario__create_image):
+            self, mock_clients, mock_cleanup):
 
         mock_clients.glance.images.get.return_value = mock.MagicMock(
             is_public=True)
 
         ctx = self.existing_image_context
         sahara_ctx = sahara_image.SaharaImage(ctx)
+        sahara_ctx._create_image = mock.Mock()
 
         sahara_ctx.setup()
         for tenant_id in sahara_ctx.context["tenants"]:
@@ -155,7 +153,7 @@ class SaharaImageTestCase(test.ScenarioTestCase):
                 sahara_ctx.context["tenants"][tenant_id]["sahara"]["image"])
             self.assertEqual("some_id", image_id)
 
-        self.assertFalse(mock_glance_scenario__create_image.called)
+        self.assertFalse(sahara_ctx._create_image.called)
 
         sahara_ctx.cleanup()
         self.assertFalse(mock_cleanup.called)